requirements.txt
alembic==0.8.6
amqp==1.4.9
anyjson==0.3.3
apipkg==1.4
arrow==0.7.0
billiard==3.3.0.23
blinker==1.4
boto==2.40.0
boto3==1.3.1
botocore==1.4.24
celery==3.1.23
cffi==1.6.0
click==6.6
coverage==4.1
cryptography==1.3.2
decorator==4.0.9
docutils==0.12
ecdsa==0.13
enum34==1.1.6
execnet==1.4.1
Fabric==1.11.1
-e git+https://github.com/pythonforfacebook/facebook-sdk.git@c01be26406f8acd146363be168e4469ba6cf0a8a#egg=facebook_sdk
fake-factory==0.5.7
Faker==0.1.4
Flask==0.11
Flask-Assets==0.11
Flask-Breadcrumbs==0.3.0
Flask-CeleryExt==0.2.0
Flask-DebugToolbar==0.10.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Menu==0.5.0
Flask-Migrate==1.8.0
Flask-Principal==0.4.0
Flask-Redis==0.1.0
Flask-S3==0.2.15
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-Session==0.2.3
Flask-SQLAlchemy==2.1
Flask-SSLify==0.1.5
Flask-WTF==0.12
futures==3.0.5
geoip2==2.3.0
hashids==1.1.0
httplib2==0.9.2
idna==2.1
ipaddr==2.1.11
ipaddress==1.0.16
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
kombu==3.0.35
Mako==1.0.4
MarkupSafe==0.23
maxminddb==1.2.0
meld3==1.0.2
MySQL-python==1.2.5
ndg-httpsclient==0.4.0
paramiko==2.0.0
passlib==1.6.5
pbr==1.10.0
Pillow==3.2.0
py==1.4.31
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyflare==1.1.5
pygeoip==0.3.2
pymongo==3.2.2
pyOpenSSL==16.0.0
pytest==2.9.2
pytest-cov==2.2.1
pytest-flask==0.10.0
pytest-xdist==1.14
python-dateutil==2.5.3
python-digitalocean==1.8.1
python-editor==1.0
python-magic==0.4.11
pytz==2016.4
PyYAML==3.11
redis==2.10.5
requests==2.10.0
requests-cache==0.4.12
selenium==2.53.2
six==1.10.0
SQLAlchemy==1.0.13
sqlalchemy-migrate==0.10.0
SQLAlchemy-Utils==0.32.7
sqlparse==0.1.19
supervisor==3.3.0
Tempita==0.5.2
timezones==1.9.6
twilio==5.4.0
-e git+https://github.com/unifispot/unifi-api/@8b8a9f9e44873a9363af5b05d5a0834e8f60fc48#egg=unifi
uuid==1.30
uWSGI==2.0.13.1
vine==1.1.0
webassets==0.11.1
Werkzeug==0.11.10
WTForms==2.1
